requirements.txt
openai-whisper
python-dotenv
pandas==1.5.3
twilio==7.16.5
openai
flask==2.2.3
python-Levenshtein==0.20.9
redis==4.5.2
SQLAlchemy==1.4.47
langchain==0.0.123
gpt-index==0.4.40
llama-index==0.4.37
mysqlclient==2.1.1
tiktoken==0.3.1
gunicorn==20.1.0
black==23.3.0
pydub==0.25.1
boto3==1.26.109
langdetect==1.0.9
nbconvert==7.4.0
jupyter==1.0.0